Add the concept of signatures and signing keys
Previously we just had a keypair for encrypting data, however moving
forward signing will be very important. Take the case where we create an
entry, we _must_ sign this entry with a key that only we have otherwise
any user could spoof creations.

Generally anything that is "owned" by someone will be signed, so for now
that makes users (we sign the entire user data so that only we could
have set any of them, otherwise users could spoof our username/etc) and
entries.

Signatures will always be validated, and an invalid signature will be
considered as a malicious trust error (as any peer can validate the
signature of a payload before distributing it)