Make small commits they said.
This is a huge amount of code which completely refactors the way iris
works at its core. Iris now operates more like bitcoin, in the idea that
it has a core block-chain for each "shard". This makes sure that every
entry posted is histroically commited.

This is mostly WIP as there is no mining, or syncing yet done for the
block chain (thats next).

Iris has some weird behaviors that go against the normal block-chain
style, namely:

- Iris considers it fine to not have every single block in our chain.
This is a hard requirement due to the fact it is in theory possible to
have a large data-set than is resonablly syncable to each new node
- Theres no money lol

<3