Refactor and implement the basics of blockchain based sharing
This commit has a lot of general fixes from the post-apocalyptic
situation that was my previous block-chain refactor. Right now we can
get two clients to kind-of-sorta sync some shards together with the
chain vaguely involved.

The next step is to properly implement content sharing and the block
chain mining/distributing.